Rename helpers to extensions
